requirements.txt
python==3.10.0
pyqt6-tools==6.4.2.3.3
pyinstaller==6.12.0